refactor: unify jsonify implementations
