+++ packages/common/log/tsconfig.json
@@ -13,6 +13,9 @@
   "references": [
     {
       "path": "../node-std"
+    },
+    {
+      "path": "../util"
     }
   ]
 }
