+++ packages/sdk/client-services/src/packlets/logging/logging-service.ts
@@ -6,6 +6,7 @@ import { Event } from '@dxos/async';
 import { Stream } from '@dxos/codec-protobuf';
 import { log, LogEntry as NaturalLogEntry, LogProcessor, shouldLog } from '@dxos/log';
 import { LoggingService, LogEntry, QueryLogsRequest } from '@dxos/protocols/proto/dxos/client/services';
+import { jsonify } from '@dxos/util';
 
 /**
  * Logging service used to spy on logs of the host.
@@ -52,32 +53,3 @@ export class LoggingServiceImpl implements LoggingService {
     this._logs.emit(entry);
   };
 }
-
-/**
- * Recursively converts an object into a JSON-compatible object.
- */
-const jsonify = (value: any, handles = new WeakSet<any>()): any => {
-  if (typeof value === 'function') {
-    return null;
-  } else if (typeof value === 'object' && value !== null) {
-    if (handles.has(value)) {
-      return null;
-    }
-    handles.add(value);
-
-    if (Array.isArray(value)) {
-      return value.map((x) => jsonify(x, handles));
-    } else {
-      if (typeof value.toJSON === 'function') {
-        return value.toJSON();
-      }
-      const res: any = {};
-      for (const key of Object.keys(value)) {
-        res[key] = jsonify(value[key], handles);
-      }
-      return res;
-    }
-  } else {
-    return value;
-  }
-};
